Update vortex.html
I was able to put five sliders on the .html and move the number read outs to the right side. I noticed, though, that it did change the overall layout of the page and I can't figure out why. 

I was thinking of maybe working on getting labels for each slider on the left side sometime this week as well but wanted to wait and see your thoughts about what's going on with the layout