add harness and generalized functions
Harness allows any member variable of a functor to include a vector of callable generalized functions. Harness () function overload updates its value, while * overload returns value.
Generalized functions are in the format class a( a, context ) though they could take more arguments. This means, for instance in the Jiggy Curly example, `struct wiggle` (an oscillator) can be re-used.
Tinkertoy functions allow a generalized function to change an particular element member, or to be parameterized by anything in the context. Functions of time allow the cluster to be animated.